+++ tvsd/__main__.py
@@ -1,9 +1,8 @@
+"""
+TVSD app entry point 
+"""
+
 from tvsd import cli
 
-
-def main():
-    cli.app(prog_name="tvsd")
-
-
-if __name__ == "__main__":
-    main()
+# if __name__ == "__main__":
+cli.app(prog_name="tvsd")