+++ tvsd/download.py
@@ -1,18 +1,16 @@
+"""TVSD Download Class"""
 import logging
 import os
 import shutil
-from typing import Literal
+from typing import Union
 
 import m3u8_To_MP4
 import typer
 
 from tvsd._variables import state_base_path, state_temp_base_path
-from multipledispatch import dispatch
-
-
-from tvsd.show import Show
-from tvsd.season import Season
-from tvsd.episode import Episode
+from tvsd.types.episode import Episode
+from tvsd.types.season import Season
+from tvsd.types.show import Show
 from tvsd.utils import mkdir_if_no
 
 
@@ -29,9 +27,10 @@ class Download:
         self._target: Show | Season | Episode = target
         self._base_path = base_path
         self._temp_base_path = temp_path
-        self._target_path: str
+
         self._specials_index: int = 1
         self._regular_ep_index: int = 1
+
         self._specials_only = specials_only
 
     def guided_download(self):
@@ -52,6 +51,7 @@ class Download:
         if download_all:
             self.download_all(self._target)
         else:
+            assert isinstance(self._target, Season)
             self.choose_download(self._target)
 
     def choose_download(self, season: "Season"):
@@ -69,24 +69,32 @@ class Download:
             else:
                 self.set_ep_index(episode)
 
-    @dispatch(Show)
-    def download_all(self, show: Show):
-        """Download all episodes in a show"""
-        logging.debug("Downloading all episodes in show")
-        for season in show.seasons:
-            self.download_all(season)
+    def download_all(self, target: Union["Season", "Show", "Episode"]):
+        """Download all episodes under the specified Season/Show/Episode"""
+        if isinstance(target, Show):
+            # Target is Show, download all seasons
+            logging.debug("Downloading all episodes in show")
+            for season in target.seasons:
+                self.download_all(season)
 
-    @dispatch(Season)
-    def download_all(self, season: Literal["Season"]):
-        """Download all episodes in a season"""
-        logging.info("Downloading all episodes in season")
+        elif isinstance(target, Season):
+            # Target is Season, download entire season
+            logging.info("Downloading all episodes in season")
 
-        # reset episode index
-        self._specials_index = 1
-        self._regular_ep_index = 1
+            # reset episode index
+            self._specials_index = 1
+            self._regular_ep_index = 1
 
-        for episode in season.episodes:
-            self.download_episode(episode)
+            for episode in target.episodes:
+                self.download_episode(episode)
+
+        elif isinstance(target, Episode):
+            # download episode
+
+            self.download_episode(target)
+
+        else:
+            raise TypeError("Target must be Show, Season or Episode")
 
     def set_special_ep_index(self, episode: "Episode"):
         """Set index for special episode
@@ -118,12 +126,6 @@ class Download:
         else:
             self.set_regular_ep_index(episode)
 
-    @dispatch(Episode)
-    def download_all(self, episode: "Episode"):
-        """Download episode"""
-
-        self.download_episode(episode)
-
     def download_episode(self, episode: "Episode"):
         """Download an episode
 
@@ -136,14 +138,21 @@ class Download:
 
         self.set_ep_index(episode)
 
+        if episode.file_exists_locally:
+            print(f"{episode.name} already exists in directory, skipping... ")
+            number = int(episode.filename.split(" - ")[1].split("E")[1])
+            if episode.is_specials:
+                self._specials_index = number + 1
+            else:
+                self._regular_ep_index = number + 1
+
+            return
+
         absolute_dest_dir = os.path.join(
             self._base_path, episode.relative_destination_dir
         )
         logging.info(absolute_dest_dir)
         mkdir_if_no(absolute_dest_dir)
-
-        if episode.file_exists_locally:
-            return
 
         print(f"Downloading to file {episode.filename}")
 
@@ -154,12 +163,15 @@ class Download:
         # m3u8_To_MP4.multithread_uri_download(m3u8_uri=episode_m3u8,
         # mp4_file_name=episode_filename, mp4_file_dir=show_dir)
         if "m3u8" not in episode_m3u8:
-            print("m3u8 Load Error, Stream probably does not exist: " + episode_m3u8)
-            return
+            logging.debug("m3u8 Load Error: %s", episode_m3u8)
+            raise ValueError(
+                "m3u8 not found in episode url, Stream probably does not exist"
+            )
 
         temp_dir = os.path.join(self._temp_base_path, episode.filename)
         if not os.path.isdir(temp_dir):
             mkdir_if_no(temp_dir)
+            print(f"Downloading {episode.filename} to {absolute_dest_dir}...")
             try:
                 m3u8_To_MP4.multithread_uri_download(
                     m3u8_uri=episode_m3u8,