+++ tvsd/sources/ssstv.py
@@ -3,7 +3,7 @@ from typing import Any
 
 from bs4 import BeautifulSoup, ResultSet, Tag
 
-from tvsd.source import Source
+from tvsd.sources.base import Source
 
 
 class SSSTV(Source):
@@ -28,46 +28,60 @@ class SSSTV(Source):
         return soup["href"]
 
     def _set_episode_title(self, soup: Tag) -> str:
-        return soup.find("span").get_text()
+        span_tag = soup.find("span")
+        if span_tag:
+            return span_tag.get_text()
+        return ""
 
     ##### PARSE SEASON FROM QUERY RESULT #####
 
     def _get_result_note(self, query_result: BeautifulSoup) -> str:
-        note = query_result.find("a", attrs={"class": "video-serial"}).get_text()
-        return note
+        note_tag = query_result.find("a", attrs={"class": "video-serial"})
+        if note_tag:
+            return note_tag.get_text()
+        return ""
 
     def _get_result_details_url(self, query_result: BeautifulSoup) -> str:
-        relative_url = query_result.find("a", attrs={"class": "video-serial"})["href"]
-
-        return f"{self._domain}{relative_url}"
+        relative_url_tag = query_result.find("a", attrs={"class": "video-serial"})
+        if relative_url_tag:
+            relative_url = relative_url_tag["href"]
+            return f"{self._domain}{relative_url}"
+        return ""
 
     #### PARSE SEASON DETAILS FROM DETAILS URL ####
 
     def _set_season_title(self, soup: BeautifulSoup):
-        return soup.find("h1", attrs={"class": "page-title"}).get_text()
+        title_tag = soup.find("h1", attrs={"class": "page-title"})
+        if title_tag:
+            return title_tag.get_text()
+        return ""
 
     def _set_season_description(self, soup: BeautifulSoup):
-        return (
-            soup.find("div", attrs={"class": "video-info-content"})
-            .find("span")
-            .get_text()
-        )
+        description_tag = soup.find("div", attrs={"class": "video-info-content"})
+        if description_tag:
+            span_tag = description_tag.find("span")
+            if span_tag is not None:
+                return span_tag.text
+        return ""
 
     def _set_season_episodes(self, soup: BeautifulSoup):
-        return (
-            soup.find("div", attrs={"class": "module-player-list"})
-            .find("div", attrs={"class": "scroll-content"})
-            .find_all("a")
-        )
+        module_player_list = soup.find("div", class_="module-player-list")
+        if module_player_list:
+            scroll_content = module_player_list.find(
+                "div", attrs={"class": "scroll-content"}
+            )
+            if scroll_content:
+                return scroll_content.find_all("a", recursive=False)
+        return []
 
     def _set_season_year(self, soup: BeautifulSoup):
-        return soup.find(
+        year_tag = soup.find(
             "a",
-            {
-                "class": "tag-link",
-                "href": re.compile(r"/vodshow/\d+-+\d{4}.html"),
-            },
-        ).get_text()
+            href=re.compile(r"/vodshow/\d+-+\d{4}.html"),
+        )
+        if year_tag:
+            return year_tag.get_text()
+        return ""
 
     ######## FETCH EPISODE M3U8 ########