+++ tvsd/utils.py
@@ -1,11 +1,20 @@
+"""
+TVSD Utilities, contains useful functions 
+"""
+
+import importlib
 import logging
+import mimetypes
 import os
+import pkgutil
 import re
+import sys
 from typing import List
+
 import cloudscraper
 import typer
-import mimetypes
-from tvsd._variables import state_base_path, state_specials_dir, state_series_dir
+
+from tvsd._variables import state_base_path, state_series_dir, state_specials_dir
 
 # from tvsd.config import BASE_PATH, SPECIALS_DIR
 
@@ -38,7 +47,8 @@ def mkdir_if_no(check_dir: str, recursive: bool = True):
                 os.mkdir(check_dir)
             except FileNotFoundError:
                 logging.error(
-                    f"Parent directory does not exist, cannot create directory {check_dir}"
+                    "Parent directory does not exist, cannot create directory %s",
+                    {check_dir},
                 )
 
 
@@ -123,7 +133,7 @@ def check_dir_mounted(path: str) -> bool:
         bool: True if directory exists
     """
 
-    if not os.path.ismount(path):
+    if not os.path.ismount(path) and not os.path.isdir(path):
         print(path, "has not been mounted yet. Exiting...")
         return False
     if not os.path.isdir(os.path.join(path, state_series_dir())):